renamed errors to noise
