+++ mkdocs.yml
@@ -40,7 +40,7 @@ nav:
   - Realistic simulations:
     - realistic_sims/index.md
     - Measurement protocols: realistic_sims/measurements.md
-    - Simulated errors: realistic_sims/errors.md
+    - Simulated errors: realistic_sims/noise.md
     - Error mitigation: realistic_sims/mitigation.md
 
   - API:
