+++ qadence/__init__.py
@@ -10,12 +10,12 @@ from .backends import *
 from .blocks import *
 from .circuit import *
 from .constructors import *
-from .errors import *
 from .exceptions import *
 from .execution import *
 from .measurements import *
 from .ml_tools import *
 from .models import *
+from .noise import *
 from .operations import *
 from .overlap import *
 from .parameters import *
@@ -44,7 +44,7 @@ list_of_submodules = [
     ".blocks",
     ".circuit",
     ".constructors",
-    ".errors",
+    ".noise",
     ".exceptions",
     ".execution",
     ".measurements",
